delete and edit field actions
