+++ apps/forms/scaffolds/blocks-editor/blocks/index.tsx
@@ -6,12 +6,20 @@ import {
   DotsHorizontalIcon,
   DragHandleHorizontalIcon,
   InputIcon,
+  Pencil1Icon,
   SectionIcon,
+  TrashIcon,
 } from "@radix-ui/react-icons";
 import React, { useCallback } from "react";
 import { useDraggable, useDroppable } from "@dnd-kit/core";
 import { CSS } from "@dnd-kit/utilities";
 import { useSortable } from "@dnd-kit/sortable";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@editor-ui/dropdown-menu";
 
 export function BlocksCanvas({
   children,
@@ -36,6 +44,8 @@ export function Block(props: React.PropsWithChildren<FormBlock>) {
     setActivatorNodeRef,
     transform,
     isDragging,
+    isSorting,
+    isOver,
     transition,
   } = useSortable({
     id: props.id,
@@ -62,13 +72,13 @@ export function Block(props: React.PropsWithChildren<FormBlock>) {
         data-folder={props.type === "section"}
         ref={setNodeRef}
         style={style}
-        className="relative data-[folder='true']:mt-16 data-[folder='true']:mb-4 data-[folder='true']:min-h-64"
+        className="relative data-[folder='true']:mt-16 data-[folder='true']:mb-4"
       >
         <button
           ref={setActivatorNodeRef}
           {...listeners}
           {...attributes}
-          className="absolute -left-4 -top-4 bg-white rounded border shadow p-1"
+          className="absolute -left-8 top-1 bg-white rounded border shadow p-1"
         >
           <DragHandleHorizontalIcon />
         </button>
@@ -96,6 +106,21 @@ export function FieldBlock({ id, type, form_field_id, data }: FormBlock) {
     [dispatch, id]
   );
 
+  const onDelete = useCallback(() => {
+    console.log("delete block", id);
+    dispatch({
+      type: "blocks/delete",
+      block_id: id,
+    });
+  }, [dispatch, id]);
+
+  const onEditClick = useCallback(() => {
+    dispatch({
+      type: "editor/field/edit",
+      field_id: form_field_id!,
+    });
+  }, [dispatch, form_field_id]);
+
   return (
     <div className="rounded-md flex flex-col gap-4 border w-full p-4 bg-white shadow-md">
       <div className="flex w-full justify-between items-center">
@@ -118,9 +143,25 @@ export function FieldBlock({ id, type, form_field_id, data }: FormBlock) {
           </select>
         </div>
         <div>
-          <button>
-            <DotsHorizontalIcon />
-          </button>
+          <DropdownMenu modal={false}>
+            <DropdownMenuTrigger asChild>
+              <button>
+                <DotsHorizontalIcon />
+              </button>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent>
+              <DropdownMenuItem onClick={onDelete}>
+                <TrashIcon />
+                Delete
+              </DropdownMenuItem>
+              {form_field_id && (
+                <DropdownMenuItem onClick={onEditClick}>
+                  <Pencil1Icon />
+                  Edit Field
+                </DropdownMenuItem>
+              )}
+            </DropdownMenuContent>
+          </DropdownMenu>
         </div>
       </div>
       <div className="w-full min-h-40 bg-neutral-200 rounded p-10 border border-black/20">
@@ -143,17 +184,25 @@ export function SectionBlock({
   children,
   ...props
 }: React.PropsWithChildren<FormBlock>) {
-  const { setNodeRef } = useDroppable({
+  const { setNodeRef, isOver } = useDroppable({
     id: props.id,
   });
 
   return (
-    <div ref={setNodeRef} className="p-4 rounded-md border bg-white shadow-md">
-      <span className="flex flex-row gap-2 items-center">
-        <SectionIcon />
-        <span>Section</span>
-      </span>
-      {children}
+    <div>
+      <div className="p-4 rounded-md border bg-white shadow-md">
+        <span className="flex flex-row gap-2 items-center">
+          <SectionIcon />
+          <span>Section</span>
+        </span>
+      </div>
+      <div
+        data-over={isOver}
+        ref={setNodeRef}
+        className="min-h-64 m-4 rounded bg-neutral-500/5 data-[over='true']:bg-blue-500/10"
+      >
+        {children}
+      </div>
     </div>
   );
 }
